unified project for SkypeWeb protocol
git-svn-id: http://svn.miranda-ng.org/main/trunk@13310 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/SkypeWeb/src/requests/avatars.h
@@ -39,7 +39,7 @@ public:
 
 		pData = (char*)mir_alloc(dataSize);
 		memcpy(pData, data, dataSize);
-		dataLength = dataSize;
+		dataLength = (int)dataSize;
 	}
 
 	~SetAvatarRequest()
